Error prevention on no previous comments found

+++ screenShotCompare.sh
@@ -43,20 +43,27 @@ if [ -z "${SCREENSHOT_PASSWORD}" ]; then
 fi
 
 echo "==> Delete all old comments, starting with 'Screenshot differs:"
-
-oldComments=$(curl_gh -X GET https://api.github.com/repos/"$GITHUB_REPOSITORY"/issues/"$PR"/comments | jq '.[] | (.id |tostring) + "|" + (.body | test("Screenshot differs:.*") | tostring)' | grep "|true" | tr -d "\"" | cut -f1 -d"|")
-echo "oldComments=$oldComments"
-echo "$oldComments" | while read comment; do
-  if [ -z "$comment" ]
-  then
-    # comment is empty
-    echo "==> old comment is empty, there is nothing to do"
-  else
-    # comment is not empty
-    echo "==> delete comment=$comment"
-    curl_gh -X DELETE https://api.github.com/repos/"$GITHUB_REPOSITORY"/issues/comments/"$comment"
-  fi
-done
+oldCommentsFirst=$(curl_gh -X GET https://api.github.com/repos/"$GITHUB_REPOSITORY"/issues/"$PR"/comments | jq '.[] | (.id |tostring) + "|" + (.body | test("Screenshot differs:.*") | tostring)')
+echo "oldCommentsFirst=$oldCommentsFirst"
+if [ -z "$oldCommentsFirst" ]
+then
+  # comment is empty
+  echo "==> oldCommentsFirst is empty, there is nothing to do"
+else
+  oldComments=$(curl_gh -X GET https://api.github.com/repos/"$GITHUB_REPOSITORY"/issues/"$PR"/comments | jq '.[] | (.id |tostring) + "|" + (.body | test("Screenshot differs:.*") | tostring)' | grep "|true" | tr -d "\"" | cut -f1 -d"|")
+  echo "oldComments=$oldComments"
+  echo "$oldComments" | while read comment; do
+    if [ -z "$comment" ]
+    then
+      # comment is empty
+      echo "==> old comment is empty, there is nothing to do"
+    else
+      # comment is not empty
+      echo "==> delete comment=$comment"
+      curl_gh -X DELETE https://api.github.com/repos/"$GITHUB_REPOSITORY"/issues/comments/"$comment"
+    fi
+  done
+fi
 
 pushd $diffFiles
 pwd
